Rich Text Editor | Add setMarkdown() and getMarkdown() methods to set and get content from Tip Tap editor  (#1424)
# Pull Request

## 🤨 Rationale
This PR contains logic for parsing the assigned markdown content,
setting it to the editor and also serializing the editor's data to
markdown content.

AzDo Feature:
https://dev.azure.com/ni/DevCentral/_backlogs/backlog/ASW%20SystemLink%20LIMS/Features/?workitem=2350963
Issue: https://github.com/ni/nimble/issues/1288

## 👩‍💻 Implementation

* Exposed getMarkdown() and setmarkdown() methods to facilitate the
exchange of data with the editor.
* Used MarkdownParser from
[prosemirror-markdown](https://github.com/ProseMirror/prosemirror-markdown/tree/master)
package for parsing markdown strings, DOM serializer from
[prosemirror-model ](https://github.com/ProseMirror/prosemirror-model)
package to serialize the content as DOM structure, XML Serializer to
serialize it to HTML string and sets it to the editor.
* To serialize the tip-tap [document
](https://prosemirror.net/docs/ref/#model.Document_Structure)in
getMarkdown() used MarkdownSerializer from
[prosemirror-markdown](https://github.com/ProseMirror/prosemirror-markdown/tree/master)
package to serialize the node based on schema.
* Enabled `emphasis` and `list` rules in MarkdownParser, this allows
users to set bold, italics, numbered, and bulleted lists in a CommonMark
flavor to the editor. All other basic markdown formats are disabled.
* Added only respective nodes and marks for bold, italics, numbered and
bulleted lists in `MarkdownSerializer`

## 🧪 Testing

* Added unit tests and visual tests for the component.
* Manually tested and verified the functionality of the supported
features.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Signed-off-by: Aagash Raaj <aagash.maridas@solitontech.com>
Co-authored-by: SOLITONTECH\vivin.krishna <123377523+vivinkrishna-ni@users.noreply.github.com>
Co-authored-by: SOLITONTECH\aagash.maridas <aagash.maridas@solitontech.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>