Add tech overview video link to CONTRIBUTING.md (#1449)
# Pull Request

## 🤨 Rationale

We recently recorded a Nimble architecture overview that might be useful
to other contributors so we want to include it in our documentation.

## 👩‍💻 Implementation

Added an NI internal link to the root CONTRIBUTING.md

While I was there I noticed that the node.js version mentioned was out
of date, so I updated it.

## 🧪 Testing

Verified that the link opens.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.