Implement 'alignment' attribute on number-text column (#1444)
# Pull Request

## 🤨 Rationale

This is part of #1011. It adds an `alignment` attribute to the
`nimble-table-column-number-text` element that can be set to `right`,
`left`, or `default` (defined in a new `NumberTextAlignment` enum). When
`alignment` is set to `default`, the alignment is determined by the
column's `format` property.

## 👩‍💻 Implementation

- Added `alignment` property to the base text cell view that defaults to
`left` so that the base text cell view can control the alignment
- Added new `NumberTextAlignment` enum with values of `right`, `left`,
and `default`
- Added `alignment` attribute & property to the
`nimble-table-column-number-text`

## 🧪 Testing

- New unit tests
- Extended storybook story and manually verified the changes look
correct
- Added additional matrix tests to cover each `format` option being
configured with each `alignment` option

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: m-akinc <7282195+m-akinc@users.noreply.github.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>