Number text decimal format (#1457)
# Pull Request

## 🤨 Rationale

This PR is part of #1011. It adds support for making a table column have
`decimal` formatting, which includes:

- Adding a new `format` value of `decimal`
- Adding `decimal-digits` attribute
- Adding validation that `decimal-digits` is in the allowed range
- Making the default alignment of a `decimal` column `right`

## 👩‍💻 Implementation

The general approach of these new capabilities is aligned with other
columns that have been implemented. A few notes that might be of
interest:

- This PR adds a `TableColumnNumberTextValidator` to validate the
`decimal-digits`. The column is invalid if its `format` is `decimal` and
`decimal-digits` is outside of the supported range.
- Added a new `DecimalFormatter` class that extends `NumberFormatter`.
It differs slightly from the `RoundToIntegerFormatter` and
`DecimalFormatter` in that the `formatter` is not static because the
number of decimals being rendered can vary between columns.

## 🧪 Testing

- New unit tests
- Added additional functionality to storybook story & manually verified
it was correct
- Added another column to the matrix tests to show the decimal format

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.