Make DateText columnConfig `undefined` when the column is invalid (#1461)
# Pull Request

## 🤨 Rationale

After a brief team discussion, we decided that it would make more sense
to make the `columnConfig` `undefined` when the date-text column was
invalid rather than having the `columnConfig` contain invalid options
(i.e. an `undefined` formatter`). This is consistent with the
number-text column.

## 👩‍💻 Implementation

Set `columnConfig` to `undefined` when creating the `DateTimeFormat`
object throws an exception rather than creating a `columnConfig` object
with an `undefined` `formatter`.

## 🧪 Testing

- Ran existing unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.