Angular support for number-text table column (#1468)
# Pull Request

## 🤨 Rationale

This is part of #1011.

It adds Angular support for the number-text column.

## 👩‍💻 Implementation

This PR adds Angular support for the number-text column. The column
isn't complete yet in `nimble-components`, but this adds Angular support
for what is implemented so far.

## 🧪 Testing

- Verified the column works as expected in the Angular example app.
- Unit tests for the directive.
- Used locally built package to verify the column can be leveraged in
SLE

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: m-akinc <7282195+m-akinc@users.noreply.github.com>