Update PR template to require justification for not auto testing (#1465)
# Pull Request

## 🤨 Rationale

When a PR doesn't include auto tests I would like to see justification
of why.

## 👩‍💻 Implementation

Updated the Testing section of the PR description template to request
this information.

## 🧪 Testing

Ugh, do I really have to justify not writing auto tests for a template
change? What a waste of my time!

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.