Spec update to include absolute URL support for rich text editor (#1394)
# Pull Request

## 🤨 Rationale

As a requirement for the comments feature, we need to support absolute
URL links in both the `nimble-rich-text-editor` and
`nimble-rich-text-viewer` components. Therefore, I am initiating this PR
to add the necessary support for absolute URL links in the editor and
viewer.

Related discussion on link support:
https://github.com/ni/nimble/pull/1369#discussion_r1272857724

## 👩‍💻 Implementation

Updated spec to add information on the design for supporting absolute
URL links in a rich text editor and viewer components.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation.
-->

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Signed-off-by: Vignesh Annamalai Sekar <vignesh.annamalai@solitontech.local>
Signed-off-by: Vignesh A S <vignesh.annamalai@ni.com>
Co-authored-by: Vignesh Annamalai Sekar <vignesh.annamalai@solitontech.local>
Co-authored-by: Vignesh A S <vignesh.annamalai@ni.com>
Co-authored-by: vikisekarNI <94439533+vikisekarNI@users.noreply.github.com>