Fix bug in number-text column where -0 is sometimes displayed (#1483)
# Pull Request

## 🤨 Rationale

There are some scenarios in the number-text column when using the
`decimal` format where -0 can be displayed even though we attempt to
avoid that. Specifically, when the number does not exactly equal -0 but
when rounded to show `decimal-digits` decimal places, only zeros will be
rendered, then the number will show as `-0.00`.

## 👩‍💻 Implementation

I couldn't find a good way to look at a number and see that it would
render only zeros when rounded to `decimal-digits` decimal places.
Therefore, the approach I came up with was multiplying the number by
10^decimal-digits and then rounding. If that number is 0, then that
means only zeros will be rendered in the number.

## 🧪 Testing

Manually tested
Added new unit test for this case

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.