Fix typo causing patterns docs to remain visible on nimble.ni.dev/storybook (#1495)
# Pull Request

## 🤨 Rationale

After #1490 I noticed that the MDX file for the "patterns" doc section
that we intended to hide was still visible in the publicly hosted
https://nimble.ni.dev/storybook/?path=/docs/patterns-docs--docs

## 👩‍💻 Implementation

The filter was excluding "patterns/" but the MDX file was titled
"Patterns/". I made it lowercase to match the filter and the folder that
contains the patterns.

## 🧪 Testing

Locally modified filter to remove check for nimble.ni.dev and verified
that the section is now hidden.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.