Create and leverage shared doc string for column operation behavior (#1493)
# Pull Request

## 🤨 Rationale

This was an action item from
https://github.com/ni/nimble/pull/1483#discussion_r1316503913

We should ensure that we have consistent docs for how grouping behaves
for columns that display formatted values in their cells. This also
applies to sorting.

## 👩‍💻 Implementation

Exported a new string from `table-column-stories-utils.ts` that explains
that grouping & sorting is done on the data field value.

## 🧪 Testing

Verified the modified stories looked correct.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: m-akinc <7282195+m-akinc@users.noreply.github.com>