Update links to W3C ARIA patterns (#1497)
# Pull Request

## 🤨 Rationale

I noticed that most of our W3C links were pointing to pages that had
been moved. This isn't ideal, so I updated the links to point to the new
location of the W3C docs.

## 👩‍💻 Implementation

Update URLs

## 🧪 Testing

Manually verified that the new links point to the appropriate pages.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>