Remove 'roundToInteger' format from the number-text column (#1525)
# Pull Request

## 🤨 Rationale

We decided to remove the `roundToInteger` format from the number-text
column because it is identical to using `decimal` format with
`decimalDigits` set to 0, and we didn't want to have to maintain two
different ways of accomplishing the same thing.

## 👩‍💻 Implementation

- Remove `roundToInteger` format from code, types, tests, storybook,
etc.

## 🧪 Testing

Manually inspected storybook to make sure I couldn't find any more
references to `roundToInteger` or `round-to-integer`

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.