Remove fixture parent (#1524)
# Pull Request

## 🤨 Rationale

We decided that it wasn't a good idea to allow a test's fixture to set a
custom parent element, and instead, the parent should explicitly be part
of the fixture by putting it in the template. See discussion here:
https://github.com/ni/nimble/pull/1499#discussion_r1322182443

## 👩‍💻 Implementation

- Update the `nimble-table-column-date-text` tests to include the
`nimble-theme-provider` within the fixture's template rather than
setting it as the parent of the fixture.
- Remove `parent` option on our fixtures to disallow setting a parent
that isn't an element within the template
- Fix typo in `nimble-table-column-number-text` tests

## 🧪 Testing

Ran the unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.