Remove unexpected whitespace at start/end of anchor (#1532)
## 🤨 Rationale

Fixes #1508

## 👩‍💻 Implementation

The issue is that whitespace inside the template was being rendered.

Inlined `startSlotTemplate` so that we could remove whitespace internal
to it. (We had previously done the same thing to the `endSlotTemplate`
for the same reason.) Removed all other whitespace in the template that
was causing a problem.

## 🧪 Testing

Tested in Storybook. Created a new Chromatic test.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.