Updated parameterized test pattern (#1379)
# Pull Request

## 🤨 Rationale

Updates the parametrized test pattern to address a few shortcomings in
the previous pattern:
- Avoids a lot of duplicated boilerplate
- Doesn't encourage an approach that results in eslint disables being
used often (false positive of await used in a loop)
- Prevents accidentally focusing a test and checking it in
- Gives decent typing, may take run-time to see if it needs improvements

## 👩‍💻 Implementation

Exposed new parameterization helpers named `parameterize` and
`parameterizeNamedList` and marked the existing method deprecated.
Created a tech debt task to track migrating to the new pattern:
https://github.com/ni/nimble/issues/1551

## 🧪 Testing

Added some tests for the new helpers and migrated a couple of existing
tests to the new pattern.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed. Added detailed comments to the helper
like the deprecated helper did.