Fix bug in table when reordering columns (#1543)
# Pull Request

## 🤨 Rationale

Resolves #1536 

## 👩‍💻 Implementation

The problem that was occurring with the bug is that `cellState` and
`cellViewTemplate` binding on the `nimble-table-row` were happening one
after another. Because of this the `cellState`, which includes the
`cellRecord` and `columnConfig` could get out-of-sync from what the
`cellViewTemplate` expected when the order of columns changes. For
example, the `cellViewTemplate` for a date-text column could get passed
a `cellRecord` and `columnConfig` for the number-text column that used
to exist at a given column index.

To solve the problem, I decided the best approach was to have the row's
template use a `repeat` directive over the `columns` with `recycle:
false`. This means that if the set of columns changes, we won't attempt
to reuse `nimble-table-cell` instances for columns that now may be
changing types. In order to facilitate this change, I could no longer
combine all state that cells need to know about into a single array of
type `ColumnState`. Now, there are parallel arrays for `columns`,
`cellIndentLevels`, and `cellStates`, which allows the row to
distinguish between the cases when the whole cell should be recreated
(i.e. the column has been changed) and when something about the cell
should be updated (i.e. the indention, configuration, or value has been
changed).

## 🧪 Testing

- Manually tested that swapping a number-text column and a date-text
column in the DOM doesn't throw an exception anymore
- Unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.