Icon table with metadata (#1554)
# Pull Request

## 🤨 Rationale

Create a table for icons in storybook that shows the tag names and
metaphors.

## 👩‍💻 Implementation

- Rendered the icons using the nimble-table
- Show metaphors in a separate column that is not sortable
- Set the height of the table large enough so that all the rows render
in the DOM. That way the page is searchable with normal ctrl+f search
and there are no nested scrollbars. Can also use normal browser zoom to
see the icons bigger.

## 🧪 Testing

Manually checked in storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.