Make tables in the Blazor and Angular example apps consistent with each other (#1555)
# Pull Request

## 🤨 Rationale

With #1519, we can re-introduce initial rows to the table in the Angular
example app without failing the Lighthouse accessibility check. I
decided that this was also a good chance to make the tables in the
Blazor and Angular example apps consistent with each other.

Resolves #1195 
Resolves #978 

## 👩‍💻 Implementation

Updated the Blazor and Angular example tables to:
- Use consistent models to represent the table data
- As part of this, I made the Blazor example app have deterministic data
and removed the use of `Faker`
- Have the same set of columns and column configuration
- Load initially with 10 rows
- Add 10 more rows to the table when the _Add rows_ button is pressed

## 🧪 Testing

- Ran both example apps and make sure the tables behave identically and
have the same data
- Verified that the Lighthouse accessibility check passes (part of the
pipeline)

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>