Rich text editor | Support for pasting absolute links when copied from other sources (#1559)
# Pull Request

## 🤨 Rationale

<!---
Provide some background and a description of your work.
What problem does this change solve?

Include links to issues, work items, or other discussions.
-->

When links are copied from other sources and pasted into the editor,
they initially render as plain text. However, they will be transformed
into links by adding whitespace or by pressing the Enter key.

Additionally, when hyperlinks (a link behind different text content) are
copied and pasted into the editor, only the text content of the link
(not the actual URL) is pasted by default or using Ctrl + V. When
pasting as plain text or using Ctrl + Shift + V, it still pastes the
same text content.

Fixes the following bugs:
1. [Bug
2525823](https://dev.azure.com/ni/DevCentral/_workitems/edit/2525823):
Rich text editor | Copying a link from Edge browser which is not pasting
as a link
3. [Bug
2525827](https://dev.azure.com/ni/DevCentral/_workitems/edit/2525827):
Rich text editor | Absolute links requires space or enter after pasting
the link

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as
complexity, alternative solutions, performance impact, etc.

Consider listing files with important changes or comment on them
directly in the pull request.
-->
1. By utilizing
[`transformPasted`](https://prosemirror.net/docs/ref/#view.EditorProps.transformPasted),
if the pasted content includes a link mark, the text content of that
node is replaced with the `href` value of the link mark. However, if the
URL within the `href` attribute doesn't start with `https` or `http`
protocols, then the text content is pasted as plain text within the same
node.
2. In addition to the `nimble-anchor` tag in `parseHTML`, added the
`<a>` tag. This is necessary because when copying a link from an
external source, the link will typically be enclosed within an `<a>`
tag, which helps the editor identify it as a link and render it within
an `<a>` tag.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->

1. To simulate the pasting of links into the editor, include the
`pasteHTMLToEditor` function in the page objects.
2. I have created unit tests to verify that when a valid anchor element
is pasted, it is displayed as a link. Additionally, when a hyperlink
from an external source is pasted, it remains an absolute link.
4. I have unit tests to check that when invalid links are pasted, they
should be rendered as plain text.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: vikisekarNI <94439533+vikisekarNI@users.noreply.github.com>