Docs improvements for "incubating" status (#1584)
# Pull Request

## 🤨 Rationale

We've had more conversations about what it takes to move a component out
of incubating. We want to capture that nuance in our docs.

## 👩‍💻 Implementation

Updated CONTRIBUTING and storybook docs for Incubating to mention that
feature gaps like missing framework integration are OK as long as they
won't cause breaking changes.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [X] I have updated the project documentation to reflect my changes or
determined no changes are needed.