Vertical scale and test mocks updates (#1577)
# Pull Request

## 🤨 Rationale

The HTML canvas has its origin in the [top-left
corner](https://www.w3schools.com/graphics/canvas_coordinates.asp).
When rendering the dies, we need to take this into consideration and
flip the vertical scale so that the origin location will be respected.

Another improvement is the wafer mock utilities for each module.
## 👩‍💻 Implementation

switched the ends of the scale range and updated tests.
replaced individual mocks with a utility method

## 🧪 Testing

tests are passing locally

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.