Update resolved version of @babel/traverse to fix critical vulnerability (#1604)
# Pull Request

## 🤨 Rationale

The `npm audit` step of the repo pipeline is failing due to a critical
vulnerability in `@babel/traverse`.

## 👩‍💻 Implementation

`@babel/traverse` is pulled in by Storybook and Angular. New versions of
those packages that uptake the patched version of `@babel/traverse` are
not yet available.

I performed these steps to get `package-lock.json` to uptake the new
version of the sub-dependency without changing any direct dependencies,
`package.json` files, or other sub-dependencies:
1. `npm i -D @babel/traverse@latest` to install the patched version at
the repo root. This updated `package-lock.json` to resolve just that
version, which is compatible with all the packages that pull it in.
2. Modify root `package.json` to remove the dependency that was added in
1.
3. `npm i`. This updated `package-lock.json` to no longer express the
direct dependcy but kept the resolved version at the newer patched one.

See alternate approaches in #1602 (just step 1 above) or #1603 (fixing
all vulnerabilities).

## 🧪 Testing

Relying on pipeline

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.